requirements.txt
kafka-python>=2.0.0
pyspark==3.5.0
pandas>=2.0.0
numpy>=1.24.0
psutil>=5.9.0
tabulate>=0.9.0
tqdm>=4.65.0